requirements.txt
streamlit
requests
beautifulsoup4
apscheduler
pytz
python-dotenv
pytest           # (테스트용)
